Solved the issue with storing the Vehicles in an array of lists. It now works. When parking a vehicle it loops through the array of parkingspots and checks if the currentsize value is appropriate for the vehicle you're trying to park. If so it creates a new List element inside that array index and stores the vehicle there. adding the vehicle size to the parking spots currentsize value.
I also started working on storing the vehicles in a json file. Right now, everytime you've successfulle parked a vehicle and it's been registered in the array of lists. It runs a method that serializes the current Array of Lists in the program onto the json file.

We still have to make a method that deserializes the json file into the Array of Lists when you launch the program.